# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MentorNepal API",
    "version": "1.0.0",
    "description": "API docs for MentorNepal backend"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Review": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "mentorId": {
            "type": "string"
          },
          "menteeId": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReviewCreate": {
        "type": "object",
        "required": [
          "sessionId",
          "rating"
        ],
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "content": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/reviews": {
      "post": {
        "summary": "Submit a review for a completed session",
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Session not completed or already reviewed"
          },
          "404": {
            "description": "Session not found"
          },
          "429": {
            "description": "Rate limit exceeded (5 req/hour)"
          }
        }
      }
    },
    "/api/reviews/{mentorId}": {
      "get": {
        "summary": "Get reviews for a mentor (cached)",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "mentorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Review"
                      }
                    },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "average": {
                          "type": "number"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mentor not found"
          }
        }
      }
    },
...
  },
...
}